+++ lib/EpdFont/builtinFonts/all.h
@@ -9,9 +9,11 @@
 //
 // Per-size guards:
 //   OMIT_TEENSY_FONT - excludes 8px (Teensy) reading fonts; used by env:xlarge
+//   OMIT_ITTY_BITTY_FONT  - excludes 9px (Itty Bitty) reading fonts; used by env:xlarge
 //   OMIT_TINY_FONT   - excludes 10px (Tiny) reading fonts; used by env:xlarge
 //   OMIT_SMALL_FONT  - excludes 12px (Small) reading fonts
 //   OMIT_MEDIUM_FONT - excludes 14px (Medium) reading fonts
+//   OMIT_LARGE_FONT  - excludes 16px (Large) reading fonts
 //   OMIT_XLARGE_FONT - excludes 18px (Extra Large) reading fonts; used by env:tiny
 //   OMIT_HUGE_FONT   - excludes 20px (Huge) reading fonts; used by all firmware envs except env:xlarge
 #ifdef OMIT_EMOJI_FONTS
@@ -27,6 +29,12 @@
 #include BUILTIN_READING_FONT_HEADER(bitter_8_bolditalic)
 #include BUILTIN_READING_FONT_HEADER(bitter_8_italic)
 #include BUILTIN_READING_FONT_HEADER(bitter_8_regular)
+#endif
+#ifndef OMIT_ITTY_BITTY_FONT
+#include BUILTIN_READING_FONT_HEADER(bitter_9_bold)
+#include BUILTIN_READING_FONT_HEADER(bitter_9_bolditalic)
+#include BUILTIN_READING_FONT_HEADER(bitter_9_italic)
+#include BUILTIN_READING_FONT_HEADER(bitter_9_regular)
 #endif
 #ifndef OMIT_TINY_FONT
 #include BUILTIN_READING_FONT_HEADER(bitter_10_bold)
@@ -46,10 +54,12 @@
 #include BUILTIN_READING_FONT_HEADER(bitter_14_italic)
 #include BUILTIN_READING_FONT_HEADER(bitter_14_regular)
 #endif
+#ifndef OMIT_LARGE_FONT
 #include BUILTIN_READING_FONT_HEADER(bitter_16_bold)
 #include BUILTIN_READING_FONT_HEADER(bitter_16_bolditalic)
 #include BUILTIN_READING_FONT_HEADER(bitter_16_italic)
 #include BUILTIN_READING_FONT_HEADER(bitter_16_regular)
+#endif
 #ifndef OMIT_XLARGE_FONT
 #include BUILTIN_READING_FONT_HEADER(bitter_18_bold)
 #include BUILTIN_READING_FONT_HEADER(bitter_18_bolditalic)
@@ -69,6 +79,12 @@
 #include BUILTIN_READING_FONT_HEADER(charein_8_italic)
 #include BUILTIN_READING_FONT_HEADER(charein_8_regular)
 #endif
+#ifndef OMIT_ITTY_BITTY_FONT
+#include BUILTIN_READING_FONT_HEADER(charein_9_bold)
+#include BUILTIN_READING_FONT_HEADER(charein_9_bolditalic)
+#include BUILTIN_READING_FONT_HEADER(charein_9_italic)
+#include BUILTIN_READING_FONT_HEADER(charein_9_regular)
+#endif
 #ifndef OMIT_TINY_FONT
 #include BUILTIN_READING_FONT_HEADER(charein_10_bold)
 #include BUILTIN_READING_FONT_HEADER(charein_10_bolditalic)
@@ -87,10 +103,12 @@
 #include BUILTIN_READING_FONT_HEADER(charein_14_italic)
 #include BUILTIN_READING_FONT_HEADER(charein_14_regular)
 #endif
+#ifndef OMIT_LARGE_FONT
 #include BUILTIN_READING_FONT_HEADER(charein_16_bold)
 #include BUILTIN_READING_FONT_HEADER(charein_16_bolditalic)
 #include BUILTIN_READING_FONT_HEADER(charein_16_italic)
 #include BUILTIN_READING_FONT_HEADER(charein_16_regular)
+#endif
 #ifndef OMIT_XLARGE_FONT
 #include BUILTIN_READING_FONT_HEADER(charein_18_bold)
 #include BUILTIN_READING_FONT_HEADER(charein_18_bolditalic)
@@ -110,6 +128,12 @@
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_8_italic)
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_8_regular)
 #endif
+#ifndef OMIT_ITTY_BITTY_FONT
+#include BUILTIN_READING_FONT_HEADER(lexenddeca_9_bold)
+#include BUILTIN_READING_FONT_HEADER(lexenddeca_9_bolditalic)
+#include BUILTIN_READING_FONT_HEADER(lexenddeca_9_italic)
+#include BUILTIN_READING_FONT_HEADER(lexenddeca_9_regular)
+#endif
 #ifndef OMIT_TINY_FONT
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_10_bold)
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_10_bolditalic)
@@ -128,10 +152,12 @@
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_14_italic)
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_14_regular)
 #endif
+#ifndef OMIT_LARGE_FONT
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_16_bold)
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_16_bolditalic)
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_16_italic)
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_16_regular)
+#endif
 #ifndef OMIT_XLARGE_FONT
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_18_bold)
 #include BUILTIN_READING_FONT_HEADER(lexenddeca_18_bolditalic)