+++ lib/EpdFont/scripts/convert-builtin-fonts.sh
@@ -127,7 +127,7 @@ PHM_FALLBACK_RANGES=(
   0x91CA,0x91CA
 )
 
-READING_FONT_SIZES=(8 10 12 14 16 18 20)
+READING_FONT_SIZES=(8 9 10 12 14 16 18 20)
 READING_FONT_STYLES=("Regular" "Bold" "Italic" "BoldItalic")
 READING_FONT_RENDER_ARGS=(--2bit --compress --pnum --darken-aa)